refactor(tv): extract shared KeyButton (size-parameterized) from SearchKeyboard
