+++ apps/tv/src/components/search/KeyButton.tsx
@@ -1,0 +1,103 @@
+import Ionicons from "@expo/vector-icons/Ionicons"
+import { useMemo } from "react"
+import { Animated, Pressable, StyleSheet, Text } from "react-native"
+
+import { scale } from "../../lib/scale"
+import { focusTransform, useFocusAnimation } from "../watch/useFocusAnimation"
+import type { KeyCell, KeyDims } from "./keyGrid"
+import { SEARCH_THEME } from "./searchTheme"
+
+type Props = {
+  cell: KeyCell
+  hasTVPreferredFocus: boolean
+  onPress: () => void
+  /** Per-keyboard size tokens (GRID_KEY_DIMS or LINEAR_KEY_DIMS). */
+  dims: KeyDims
+}
+
+/**
+ * One focusable, animated keyboard key, shared by the grid (Android) and
+ * single-line (Apple TV) keyboards. Sizing comes entirely from `dims` so the
+ * two keyboards differ only in numbers, not behavior:
+ *
+ *   - focus pop via useFocusAnimation → focusTransform (lift 0, magnify 1.1);
+ *     the timing stops the prior animation before the next so a rapid D-pad
+ *     sweep can't orphan animations;
+ *   - white-fill focus (SEARCH_THEME.keyFocusBg) + near-black ink;
+ *   - the backspace key renders an Ionicons glyph instead of a text label.
+ */
+export function KeyButton({ cell, hasTVPreferredFocus, onPress, dims }: Props) {
+  const { focused, setFocused, progress } = useFocusAnimation()
+
+  const keyTransform = useMemo(
+    () => focusTransform(progress, { lift: 0, magnify: 1.1 }),
+    [progress],
+  )
+
+  const inkColor = focused ? SEARCH_THEME.keyFocusText : SEARCH_THEME.keyText
+
+  return (
+    <Pressable
+      onPress={onPress}
+      onFocus={() => setFocused(true)}
+      onBlur={() => setFocused(false)}
+      hasTVPreferredFocus={hasTVPreferredFocus}
+      accessibilityRole="button"
+      accessibilityLabel={cell.accessibilityLabel ?? cell.label}
+    >
+      <Animated.View
+        style={[
+          styles.key,
+          {
+            width: scale(cell.wide === true ? dims.wideWidth : dims.size),
+            height: scale(dims.size),
+            borderRadius: scale(dims.radius),
+          },
+          focused && styles.keyFocused,
+          { transform: keyTransform },
+        ]}
+      >
+        {cell.action.kind === "backspace" ? (
+          <Ionicons
+            name="backspace-outline"
+            size={scale(dims.iconSize)}
+            color={inkColor}
+          />
+        ) : (
+          <Text
+            style={[
+              styles.keyLabel,
+              {
+                color: inkColor,
+                fontSize: Math.round(scale(dims.labelFontSize)),
+              },
+            ]}
+          >
+            {cell.label}
+          </Text>
+        )}
+      </Animated.View>
+    </Pressable>
+  )
+}
+
+const styles = StyleSheet.create({
+  key: {
+    backgroundColor: SEARCH_THEME.keyBg,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  keyFocused: {
+    backgroundColor: SEARCH_THEME.keyFocusBg,
+    // Design: 0 12px 28px -10px rgba(0,0,0,.7).
+    shadowColor: "#000000",
+    shadowOffset: { width: 0, height: scale(12) },
+    shadowRadius: scale(14),
+    shadowOpacity: 0.7,
+    elevation: 8,
+  },
+  keyLabel: {
+    fontFamily: "System",
+    fontWeight: "600",
+  },
+})
